规则检测去掉超时40分钟限制 规则日志数据类型匹配修改

+++ asset-collect/src/main/java/com/ruoyi/shareproject/hmuserst/domain/TLjRuleUserLogs.java
@@ -270,11 +270,11 @@ public class TLjRuleUserLogs extends BaseEntity
 
     /** 案件处理率条件1 */
     @Excel(name = "案件处理率条件1")
-    private Integer caseRateOne;
+    private BigDecimal caseRateOne;
 
     /** 案件处理率条件2 */
     @Excel(name = "案件处理率条件2")
-    private Integer caseRateTwo;
+    private BigDecimal caseRateTwo;
 
     /** 实际还款条件 */
     @Excel(name = "实际还款条件")
@@ -282,11 +282,11 @@ public class TLjRuleUserLogs extends BaseEntity
 
     /** 实际还款条件1 */
     @Excel(name = "实际还款条件1")
-    private Integer sjRepaymentOne;
+    private BigDecimal sjRepaymentOne;
 
     /** 实际还款条件2 */
     @Excel(name = "实际还款条件2")
-    private Integer sjRepaymentTwo;
+    private BigDecimal sjRepaymentTwo;
 
     /** 实际还款率条件 */
     @Excel(name = "实际还款率条件")
@@ -294,9 +294,9 @@ public class TLjRuleUserLogs extends BaseEntity
 
     /** 实际还款率条件1 */
     @Excel(name = "实际还款率条件1")
-    private Integer sjRepaymentRateOne;
+    private BigDecimal sjRepaymentRateOne;
 
     /** 实际还款率条件2 */
     @Excel(name = "实际还款率条件2")
-    private Integer sjRepaymentRateTwo;
+    private BigDecimal sjRepaymentRateTwo;
 }
